tests/wpt/tests/html/canvas/element/fill-and-stroke-styles/2d.strokeStyle.relativecolor.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.strokeStyle.relativecolor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.strokeStyle.relativecolor</h1>
<p class="desc">Relative color works as color input</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Relative color works as color input");
_addTest(function(canvas, ctx) {

  ctx.strokeStyle = "rgb(from red g r b)";
  _assertSame(ctx.strokeStyle, 'color(srgb 0 1 0)', "ctx.strokeStyle", "'color(srgb 0 1 0)'");
  ctx.strokeStyle = "color(from color(srgb 0.25 0.5 0.75 / 0.5) srgb r g b / alpha)";
  _assertSame(ctx.strokeStyle, 'color(srgb 0.25 0.5 0.75 / 0.5)', "ctx.strokeStyle", "'color(srgb 0.25 0.5 0.75 / 0.5)'");

});
</script>
